+++ docsig/__init__.py
@@ -1,5 +1,6 @@
 """Check signature params for proper documentation."""
+from ._core import docsig
 from ._main import main
 from ._version import __version__
 
-__all__ = ["__version__", "main"]
+__all__ = ["__version__", "docsig", "main"]