+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 [Unreleased](https://github.com/jshwi/docsig/compare/v0.15.0...HEAD)
 ------------------------------------------------------------------------
 ### Added
+- Adds `docsig` function for API usage
 - Adds default path argument as current working dir
 
 ### Fixed